More bug crunching.

+++ include/simdvariablebyte.h
@@ -27,9 +27,9 @@ size_t masked_vbyte_read_loop_fromcompressedsize_delta(const uint8_t* in, uint32
 		size_t inputsize, uint32_t  prev);
 //size_t read_ints(const uint8_t* in, uint32_t* out, int length) ;
 //size_t read_ints_delta(const uint8_t* in, uint32_t* out, int length, uint32_t  prev) ;
-uint32_t masked_vbyte_select_delta(const uint8_t *in, int length,
-                uint32_t prev, int slot);
-int masked_vbyte_search_delta(const uint8_t *in, int length, uint32_t prev,
+uint32_t masked_vbyte_select_delta(const uint8_t *in, uint64_t length,
+                uint32_t prev, size_t slot);
+int masked_vbyte_search_delta(const uint8_t *in, uint64_t length, uint32_t prev,
                 uint32_t key, uint32_t *presult);
 }
